+++ app/tests/e2e/test_fixtures.py
@@ -1,0 +1,93 @@
+import asyncio
+import json
+from typing import Any, Dict
+
+import pytest
+from aries_cloudcontroller import AcaPyClient
+from httpx import AsyncClient
+
+from app.admin.governance.schemas import SchemaDefinition, create_schema
+from app.dependencies import MEMBER_AGENT_URL
+from app.facades.acapy_ledger import create_pub_did
+from app.tests.utils_test import get_random_string
+from app.generic.issuer.issuer import router
+
+BASE_PATH = router.prefix + "/credentials"
+
+
+@pytest.yield_fixture(scope="module")
+def event_loop(request):
+    loop = asyncio.get_event_loop_policy().new_event_loop()
+    yield loop
+    loop.close()
+
+
+@pytest.fixture(scope="module")
+async def schema_definition(yoma_agent_module_scope: AcaPyClient) -> Dict[str, Any]:
+    definition = SchemaDefinition(
+        name="test_schema", version="0.3", attributes=["speed"]
+    )
+
+    public_did = await create_pub_did(yoma_agent_module_scope)
+    print(f"created did: {public_did}")
+
+    schema_definition_result = await create_schema(definition, yoma_agent_module_scope)
+    print(schema_definition_result)
+
+    print(f"created schema {str(schema_definition_result)}")
+    return (schema_definition_result).dict()
+
+
+@pytest.fixture(scope="module")
+async def credential_definition_id(
+    async_client_bob_module_scope: AsyncClient, schema_def: Dict[str, Any]
+) -> str:
+    # when
+    response = await async_client_bob_module_scope.post(
+        MEMBER_AGENT_URL + "/admin/governance/credential-definitions",
+        data=json.dumps(
+            {
+                "support_revocation": False,
+                "schema_id": schema_def["schema_id"],
+                "tag": get_random_string(5),
+            }
+        ),
+    )
+    result = response.json()
+
+    print(f"created definition {str(result)}")
+    return result["credential_definition_id"]
+
+
+@pytest.fixture(scope="module")
+async def credential_exchange_id(
+    async_client_bob_module_scope: AsyncClient,
+    alice_connection_id: str,
+    schema_definition: Dict[str, Any],
+    bob_connection_id: str,
+    async_client_alice_module_scope: AsyncClient,
+):
+    """this fixture produces the CRED_X_ID but if the test that produces the CRED_X_ID has already run
+    then this fixture just returns it..."""
+    credential = {
+        "protocol_version": "v1",
+        "connection_id": bob_connection_id,
+        "schema_id": schema_definition["schema_id"],
+        "attributes": {"speed": "average"},
+    }
+
+    response = await async_client_bob_module_scope.post(
+        "/generic/issuer/credentials",
+        data=json.dumps(credential),
+    )
+    credential_exchange = response.json()
+    credential_exchange_id = credential_exchange["credential_id"]
+    assert credential_exchange["protocol_version"] == "v1"
+
+    response = await async_client_alice_module_scope.get(
+        BASE_PATH, params={"connection_id": alice_connection_id}
+    )
+    records = response.json()
+    assert len(records) > 0
+
+    return credential_exchange_id
