+++ app/tests/test_dependencies.py
@@ -1,5 +1,5 @@
-from contextlib import asynccontextmanager
 import os
+from contextlib import asynccontextmanager
 
 import pytest
 from aries_cloudcontroller import AcaPyClient
