+++ app/tests/test_via_web_wallet.py
@@ -1,5 +1,5 @@
-import pytest
 import httpx
+import pytest
 from aiohttp import ClientResponseError
 from httpx import AsyncClient
