+++ app/tests/utils_test.py
@@ -1,5 +1,6 @@
 import random
 import string
+
 from httpx import AsyncClient
 
 APPLICATION_JSON_CONTENT_TYPE = {"content-type": "application/json"}
